# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nakudin Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List shops
      parameters:
        - in: query
          name: cursor
          schema: { type: ["string", "null"] }
        - in: query
          name: category
          schema: { type: ["string", "null"] }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Shop list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopPage"

    post:
      operationId: createShop
      tags: [shops]
      summary: Create a shop (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/check-name:
    get:
      operationId: checkBusinessName
      tags: [shops]
      summary: Check if a business name is available
      parameters:
        - in: query
          name: name
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Name availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameAvailability"
  /shops/me:
    get:
      operationId: getMyShop
      tags: [shops]
      summary: Get authenticated user's shop
      responses:
        "200":
          description: My shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

    patch:
      operationId: updateMyShop
      tags: [shops]
      summary: Update authenticated user's shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{shopId}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get a shop by ID
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /shops/{shopId}/follow:
    post:
      operationId: followShop
      tags: [social]
      summary: Follow a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"

    delete:
      operationId: unfollowShop
      tags: [social]
      summary: Unfollow a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"
  /shops/{shopId}/reviews:
    get:
      operationId: getShopReviews
      tags: [social]
      summary: Get reviews for a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [social]
      summary: Post a review for a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /shops/{shopId}/can-review:
    get:
      operationId: canReviewShop
      tags: [social]
      summary: Check if user can review this shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Can review status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanReviewResult"
  /shops/{shopId}/analytics:
    get:
      operationId: getShopAnalytics
      tags: [analytics]
      summary: Get analytics for a shop
      parameters:
        - in: path
          name: shopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopAnalytics"
...
components:
  schemas:
...
    Shop:
      type: object
      required: [id, businessName, category, followerCount, followingCount, verified, subscriptionStatus, createdAt]
      properties:
        id: { type: string }
        businessName: { type: string }
        bio: { type: ["string", "null"] }
        category: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        whatsappNumber: { type: ["string", "null"] }
        locationCity: { type: ["string", "null"] }
        locationState: { type: ["string", "null"] }
        locationLat: { type: ["number", "null"] }
        locationLng: { type: ["number", "null"] }
        verified: { type: boolean }
        followerCount: { type: integer }
        followingCount: { type: integer }
        totalViews: { type: integer }
        totalLikes: { type: integer }
        totalWhatsappClicks: { type: integer }
        avgRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        subscriptionStatus: { type: string, enum: [trial, active, grace, locked] }
        trialEndsAt: { type: ["string", "null"] }
        isFollowed: { type: boolean }
        createdAt: { type: string }
    ShopInput:
      type: object
      required: [businessName, whatsappNumber, category]
      properties:
        businessName: { type: string, minLength: 2 }
        whatsappNumber: { type: string }
        category: { type: string }
        bio: { type: string }
        logoUrl: { type: string }
        coverUrl: { type: string }
        locationCity: { type: string }
        locationState: { type: string }
        locationLat: { type: number }
        locationLng: { type: number }
    ShopUpdate:
      type: object
      properties:
        businessName: { type: string, minLength: 2 }
        whatsappNumber: { type: string }
        category: { type: string }
        bio: { type: string }
        logoUrl: { type: string }
        coverUrl: { type: string }
        locationCity: { type: string }
        locationState: { type: string }
        locationLat: { type: number }
        locationLng: { type: number }
    ShopPage:
      type: object
      required: [shops, nextCursor]
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        nextCursor: { type: ["string", "null"] }
    NameAvailability:
      type: object
      required: [available]
      properties:
        available: { type: boolean }
...
    FollowResult:
      type: object
      required: [followed, followerCount]
      properties:
        followed: { type: boolean }
        followerCount: { type: integer }
    CanReviewResult:
      type: object
      required: [canReview]
      properties:
        canReview: { type: boolean }
...
    Review:
      type: object
      required: [id, shopId, authorId, authorName, rating, createdAt]
      properties:
        id: { type: string }
        shopId: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        authorLogoUrl: { type: ["string", "null"] }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: ["string", "null"] }
        ownerReply: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string }
...
    ShopAnalytics:
      type: object
      required: [shopId, totalViews, totalLikes, totalWhatsappClicks, followerCount, recentActivity]
      properties:
        shopId: { type: string }
        totalViews: { type: integer }
        totalLikes: { type: integer }
        totalWhatsappClicks: { type: integer }
        followerCount: { type: integer }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityPoint"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/ProductStat"
    ActivityPoint:
      type: object
      required: [date, views, likes, clicks]
      properties:
        date: { type: string }
        views: { type: integer }
        likes: { type: integer }
        clicks: { type: integer }
    ProductStat:
      type: object
      required: [productId, title, views, likes, clicks]
      properties:
        productId: { type: string }
        title: { type: string }
        views: { type: integer }
        likes: { type: integer }
        clicks: { type: integer }
...