# lib/api-zod/src/generated/types/shopInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ShopInput {
  /** @minLength 2 */
  businessName: string;
  whatsappNumber: string;
  category: string;
  bio?: string;
  logoUrl?: string;
  coverUrl?: string;
  locationCity?: string;
  locationState?: string;
  locationLat?: number;
  locationLng?: number;
}